Soften language in why_zig_rust_d_cpp.en.md
Based on the rest of the text I guess this was meant to say that Zig intends to ship with a package manager rather than that it already ships with one. But this is a guess on my part so feel free to reject.

The context for this is https://news.ycombinator.com/item?id=34136704.

Btw how do updates to the translated articles get made?