Small typos correction and a small change to make it more understandable for different Portuguese speakers.
There were some minor errors in the Portuguese translation, mostly some missing accentuation and also a missing proposition. But it's important to take note that Portuguese from Portugal is slightly different from Brazilian Portuguese, with this in mind, I removed a word in order to make it more explicit and more understandable for Portuguese speakers from another countries and that was about it. Hope it helps.  😄