fix download table for versions that don't have all entries
The download template needs to be able to tell when it's on the last
row of a section in order to add a border to delineate sections. However,
not every version of zig has the same number of rows.  This modifies
the template to first filter the rows before it then iterates over
them.