don't color background of OS table headers
Coloring the background of the table headers was odd because it's color
is always the color of the first row in it's section.  This means you
can sometimes have adjacent section headers with the same color and
sometimes not, which looks off.  This disables changing the background
color of the table headers and only applies it to the td elements.