downloads: dark borders for light mode
The borders that separate each OS section in the download table are
hard to see in light mode.  The color is currently shared with the
background color used for every other row (#f2f3f3).  Note that in
dark mode we use the color grey for these borders which is #888.  This
has a much higher contrast with the background than light mode.  I've
changed it to #aaa which is much more visible.

This change was a result of a user getting confused because they couldn't
see that the x86_64 linux variant was apart of the "Linux" section.