css: fixes and improvements (#403)
* css: remove unused class

Also absent in the language reference.

* fix: footer links invisible on mobile

This resulted in a broken sentence, since linkes were in the middle
of it. The no-mobile class is not used anywhere else on the website
nor the language reference.

* css: remove unnecessary class used only once in the website

* css: separate language menu from the bottom of the page

* css: move devlog-specific CSS outside global stylesheet

* css: prevent div from taking space when it has no contents

* fix: devlog entries are not aligned with page title

* css: separate headings from the previous section end a bit more

Makes the visual hierarchy slightly more clear.
In some cases a paragraph is too close to the following
heading, resulting in unclear separation between sections.

* css: improve devlog highlight boxes and posts separation

* css: remove unused classes

These are unused both in the website and in the language reference.
`.sgr` classes are used for terminal output.