Introduce community mirrors
The 'Download' page has a link to the new 'Community Mirrors' page which
describes how to use them. The mirror list is accessible at
'/download/community-mirrors.txt', which is generated from the Ziggy
source in 'assets/community-mirrors.ziggy'.

I have confirmed with the owners of these Zig mirrors that they are
happy for their mirrors to be used in this context.