Introduce CI workflow to check mirrors
Every 24 hours, fetch a bunch of tarballs from ziglang.org and every
mirror, to check that they are all working as intended. A summary is
emitted in a Markdown table which GitHub will render on the workflow run
page. If any mirror fails to serve a tarball or serves one which differs
from the one on ziglang.org, the workflow sends me (@mlugg) a push
notification so I see the issue and can address it.

As well as running on a cron schedule, this workflow can also be
manually triggered and runs when the mirror list file is edited.