check-mirrors: run on GitHub-hosted runner
Minimizing the amount of code running on the website machine is
important to avoid potential security vulnerabilities.

Since we now need to fetch Zig (rather than it being preinstalled), also
update `check-mirrors` to the latest Zig dev version. We'll want to do
this soon anyway due to ziglang/zig#24316 blocking #492.