build.zig.zon: bump minimum_zig_version
It probably makes sense for this to be a full release rather than a
patch release where possible. The actual motivation here is that
mlugg/setup-zig currently uses this version exactly. In fact, for `-dev`
versions, I don't think it even *can* do anything else, since there may
be breaking changes between `0.A.0-dev+...` and `0.A.0`.